Add validation

+++ src/main/java/by/schepov/motordepot/command/SubmitOrder.java
@@ -1,15 +1,13 @@
 package by.schepov.motordepot.command;
 
 import by.schepov.motordepot.builder.impl.order.OrderBuilder;
-import by.schepov.motordepot.entity.Car;
-import by.schepov.motordepot.entity.Order;
-import by.schepov.motordepot.entity.Request;
-import by.schepov.motordepot.entity.User;
+import by.schepov.motordepot.entity.*;
 import by.schepov.motordepot.exception.CarServiceException;
 import by.schepov.motordepot.exception.service.OrderServiceException;
 import by.schepov.motordepot.exception.service.RequestServiceException;
 import by.schepov.motordepot.jsp.JSPParameter;
 import by.schepov.motordepot.jsp.Page;
+import by.schepov.motordepot.jsp.RequestAttribute;
 import by.schepov.motordepot.service.car.CarService;
 import by.schepov.motordepot.service.car.impl.CarRepositoryService;
 import by.schepov.motordepot.service.order.OrderService;
@@ -55,10 +53,18 @@ public class SubmitOrder implements Executable{
                 LOGGER.warn("Car hasn't been found(id=" + carId + ")");
                 return Page.ERROR;
             }
+            if(foundCar.getCarStatus() == CarStatus.BUSY){
+                LOGGER.info("The car was assigned by another admin(car_id=" + carId + ")");
+                request.setAttribute(RequestAttribute.REQUEST.getName(), foundRequest);
+                Set<Car> cars = carService.findFreeCars(foundRequest.getLoad(), foundRequest.getPassengersQuantity());
+                request.setAttribute(RequestAttribute.CARS.getName(), cars);
+                return Page.REQUEST_VERIFICATION;
+            }
             orderBuilder.reset();
             Order order = orderBuilder.withUser(foundRequest.getUser()).withCar(foundCar).withDriver(foundCar.getDriver())
                     .withDepartureLocation(foundRequest.getDepartureLocation()).withArrivalLocation(foundRequest.getArrivalLocation())
                     .build();
+            carService.updateCarStatus(carId, CarStatus.BUSY);
             requestService.deleteRequest(foundRequest);
             orderService.insertOrder(order);
             return Page.MANAGEMENT;
